feat: link to courseware for existing enrollment with CourseRunCard
