+++ src/components/course/course-header/data/index.js
@@ -1,0 +1,1 @@
+export { default as getCourseRunCardProps } from './getCourseRunCardProps';
